fixed_asset_register/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==26.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
google-auth==2.27.0
requests==2.32.5
django-cors-headers>=4.3.0
pyjwt==2.8.0
